+++ CHANGELOG.md
@@ -13,8 +13,8 @@
 
 #### Improvements
 
-- Improved query generation for `Dataframe.distinct` to generate `SELECT DISTINCT` instead of `SELECT` with `GROUP BY` all columns.
 - Improved the random object name generation to avoid collisions.
+- Improved query generation for `Dataframe.distinct` to generate `SELECT DISTINCT` instead of `SELECT` with `GROUP BY` all columns. To disable this feature, set `session.conf.set("use_simplified_query_generation", False)`.
 
 #### Deprecations